Package restructure and removal of dependency on message SDK
N.B. This is a breaking change.

It was necessary to alter the package structure in this project so that projects which
depend on this project and the message SDK do not have conflicts in class loading.

Moreover, tests in this project were dependent on message SDK classes (the dependency
was test-scoped). This has been changed - there is no dependency on the message SDK
at all.