Package restructure and removal of dependency on message SDK (#2)
* Package restructure and removal of dependency on message SDK

N.B. This is a breaking change.

It was necessary to alter the package structure in this project so that projects which
depend on this project and the message SDK do not have conflicts in class loading.

Moreover, tests in this project were dependent on message SDK classes (the dependency
was test-scoped). This has been changed - there is no dependency on the message SDK
at all.

* Removed unnecessary package information from classes in test case

* Changed version back to 1.0.0

In keeping with the proper proposed git flow, the version bump should happen on a release branch.