v2.0.0 [release-type final]
V2: Package restructure and JsonPath-based JSON masking (#5)

* Package restructure and removal of dependency on message SDK (#2)

* Package restructure and removal of dependency on message SDK

N.B. This is a breaking change.

It was necessary to alter the package structure in this project so that projects which
depend on this project and the message SDK do not have conflicts in class loading.

Moreover, tests in this project were dependent on message SDK classes (the dependency
was test-scoped). This has been changed - there is no dependency on the message SDK
at all.

* Removed unnecessary package information from classes in test case

* Changed version back to 1.0.0

In keeping with the proper proposed git flow, the version bump should happen on a release branch.

* Feature: no whitespace masking (#4)

* Added check to ensure whitespace-only strings aren't masked

Must still add a test case.

* Added test case to confirm no whitespace masking

* Feature json masking (#3)

* Initial commit

Exact package structure and classes involving masking copied from the message SDK. Build section of the POM was copied from the message SDK. 

Using circle v2 build config.

* Major.minor.patch versioning, starting at major version 1

* Bump artifactory-maven-plugin dependency

* Target Java 8

* Set version to 1.00

* Added a JSON masker utility class, comparable to XML masking.

Included test cases to verify masking behaviour.

* Pretty printed return for JSON masking

* Fixed references to removed package

* Masking methods now throw masking exceptions when null input is supplied.

Null is not valid XML/JSON, and both Utils classes currently throw exceptions
when any other masking failure occurs - calling code has to wrap calls in try/catch
blocks anyway. Calling code must decide how to handle masking failures.

* Added Javadoc to indicate null xml strings are not processed.

* Testing for valid but empty JSON/XML

* Trigger circle build with no release type

* v2.0.0-RC1

* Javadoc corrections

* Explicit null check for set of masking units

If the set of masking units to be used in JSON/XML masking is null,
throw a *MaskingException. This allows the caller to appropriately
handle failed masking.

* v2.0.0-RC2

* Draft of release notes document

* Fix bug where PANs of length ten are not masked at all

* Release notes and final version